backend/myAISearchCreator/requirements.txt
azure-core==1.29.2
python-dotenv==1.0.0
azure-search-documents==11.5.2
azure-storage-blob==12.18.0
azure-cosmos

